requirements.txt
pyzmq
pyYaml
keyboard
python-gnupg
